Added a functions
added a function that takes the raw int input from the user and changes it to a readable str for the json url file. Afterwards made made corrections in order for the code to work. It takes the input and checks the currencies, it just needs tweaks in order to make sure the user inputs an int